crud santri dan asatidz, global data

+++ app/Http/Controllers/Admin/IdentitasController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Models\Periode;
 use App\Models\Setting;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Intervention\Image\ImageManager;
 use Intervention\Image\Drivers\Gd\Driver;
+use App\Providers\GlobalDataServiceProvider;
 use App\Providers\SettingServiceProvider as SettingService;
 use App\Providers\AlertResponseServiceProvider as AlertResponse;
 
@@ -17,7 +19,9 @@ class IdentitasController extends Controller
      */
     public function index()
     {
-        $data = Setting::getAllSettingMadin();
+        $data = GlobalDataServiceProvider::get();
+        $data['setting'] = Setting::getAllSettingMadin();
+
 
         return view('admin.setting.identitas', $data);
     }
@@ -52,8 +56,10 @@ class IdentitasController extends Controller
         Setting::setSettingMadin('logo_madin', $imageName);
         $msg = AlertResponse::response('success', 'Berhasil ubah logo madin!');
 
-        if (file_exists(storage_path('app/public/logo/' . $oldImage))) {
-            unlink(storage_path('app/public/logo/' . $oldImage));
+        if ($oldImage != 'logo.png') {
+            if (file_exists(storage_path('app/public/logo/' . $oldImage))) {
+                unlink(storage_path('app/public/logo/' . $oldImage));
+            }
         }
 
         return response()->json([$msg, $imageName]);
